+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 CHANGELOG
 ==========
 
+0.0.2 Jun 4, 2012
+-----------------
+
+* Change paginator to kaminari in extras
+* Add i18n option to extras to set zh-CN locale and time zone
+
 0.0.1 Jun 3, 2012
 -----------------
 
 * Create my own thyone app creator
 * Fork from Rails_Apps_Composer
-